// frontend/src/types/user.types.ts
export interface UserType {
  bio: string;
  coverImage: string;
  createdAt: string;
  email: string;
  followers: unknown[];
  following: unknown[];
  fullName: string;
  likedPosts: unknown[];
  link: string;
  profileImage: string;
  updatedAt: string;
  userName: string;
  __v: number;
  _id: string;
}